Charts

Charts represent maps from parameter domains to configuration space. Their main use is to provide the geometric description of the cells that make up meshes.

The Chart concept is defined by the following API

CompScienceMeshes.dimension — Method

dimension(simplex)

Return the manifold dimension of the simplex.

CompScienceMeshes.universedimension — Function

universedimension(mesh)

Returns the dimension of the surrounding space. Equals the number of coordinates required to describe a vertex.

universedimension(p)

Return the dimension of the universe in which p is embedded.

CompScienceMeshes.volume — Method

volume(simplex)

Return the volume of the simplex.

CompScienceMeshes.neighborhood — Function

neighborhood(chart, params)

Create a neighborhood from a chart and a set of parameter values.

The most important example of a Chart is a Simplex. For Simplices, the following additional functions are available

CompScienceMeshes.simplex — Function

simplex(vertices)
simplex(v1, v2, ...)
simplex(vertices, Val{D})

Build a D-dimensional simplex. The vertices can be passed in an array (static or dynamic), or supplied separately. If the length of the array is not part of its type, the speed of the construction can be improved by supplying an extra Val{D} argument. In case it is not clear from the context whether the vertex array is dynamically or statically sized, use the third form as it will not incur notable performance hits.

Note that D is the dimension of the simplex, i.e. the number of vertices supplied minus one.

CompScienceMeshes.center — Function

center(simplex) -> neighborhood

Create the neighborhood at the center of the simplex, i.e. the point corresponding to parameter (1/(D+1), 1/(D+1), ...) where D is the simplex dimension.